Fix tests
Now that we use pytest-asyncio, it is easier to test the dead links in an async context and not need to worry about weird issues with the aiohttp client session being used outside of an async context. The places where the function was being called before (only one, in search) can now call `sync_check_dead_links` instead without any change to the overall behaviour of the app. It just makes the tests easier to configure now that we actually have real async contexts going on in some places.
