+++ api/test/unit/utils/test_check_dead_links.py
@@ -1,16 +1,13 @@
 import asyncio
 from unittest import mock
 
-import aiohttp
 import pook
 import pytest
 
 from api.utils.check_dead_links import HEADERS, check_dead_links
 
 
-@mock.patch.object(aiohttp, "ClientSession", wraps=aiohttp.ClientSession)
-@pook.on
-def test_sends_user_agent(wrapped_client_session: mock.AsyncMock):
+async def test_sends_user_agent(pook_on):
     query_hash = "test_sends_user_agent"
     results = [{"provider": "best_provider_ever"} for _ in range(40)]
     image_urls = [f"https://example.org/{i}" for i in range(len(results))]
@@ -18,22 +15,18 @@ def test_sends_user_agent(wrapped_client_session: mock.AsyncMock):
 
     head_mock = (
         pook.head(pook.regex(r"https://example.org/\d"))
+        .headers(HEADERS)
         .times(len(results))
         .reply(200)
         .mock
     )
 
-    check_dead_links(query_hash, start_slice, results, image_urls)
+    await check_dead_links(query_hash, start_slice, results, image_urls)
 
     assert head_mock.calls == len(results)
-    requested_urls = [req.rawurl for req in head_mock.matches]
-    for url in image_urls:
-        assert url in requested_urls
-
-    wrapped_client_session.assert_called_once_with(headers=HEADERS, timeout=mock.ANY)
 
 
-def test_handles_timeout():
+async def test_handles_timeout():
     """
     Test that case where timeout occurs.
 
@@ -51,7 +44,7 @@ def test_handles_timeout():
     with mock.patch(
         "aiohttp.client.ClientSession._request", side_effect=raise_timeout_error
     ):
-        check_dead_links(query_hash, start_slice, results, image_urls)
+        await check_dead_links(query_hash, start_slice, results, image_urls)
 
     # `check_dead_links` directly modifies the results list
     # if the results are timing out then they're considered dead and discarded
@@ -60,7 +53,7 @@ def test_handles_timeout():
 
 
 @pytest.mark.parametrize("provider", ("thingiverse", "flickr"))
-def test_403_considered_dead(provider):
+async def test_403_considered_dead(provider):
     query_hash = f"test_{provider}_403_considered_dead"
     other_provider = "fake_other_provider"
     results = [
@@ -77,7 +70,7 @@ def test_403_considered_dead(provider):
         .mock
     )
 
-    check_dead_links(query_hash, start_slice, results, image_urls)
+    await check_dead_links(query_hash, start_slice, results, image_urls)
 
     assert head_mock.calls == len(results)
